requirements-dev.txt
asgiref==3.2.10
certifi==2020.12.5
dj-database-url==0.5.0
Django==3.1
django-cors-headers==3.4.0
django-environ==0.4.5
django-templated-mail==1.1.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.6.0
djoser==2.0.3
Pillow==8.1.0
PyJWT==2.0.0
pytz==2020.5
sqlparse==0.4.1
